data download and deepchecks scripts updated, src function files added for read_zip and data_deepchecks
